Restructure files in folders (#87)
* Move drawPlayer.js into game folder and fix imports

* Move rotatePlayer.js into game folder and fix import

* Move checkIfPlayerIsByRocket.js into game folder and fix import
